Exceptions
Exceptions bad.request delete 404

+++ dscommerce/src/main/java/com/devsuperior/dscommerce/service/ProductService.java
@@ -3,17 +3,23 @@
  */
 package com.devsuperior.dscommerce.service;
 
-import java.util.Optional;
+
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.service.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.service.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 
@@ -29,8 +35,9 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) 
 	{
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
+		
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 		
@@ -52,10 +59,15 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update( Long id, ProductDTO dto){
+		try {
 		Product entity = repository.getReferenceById(id);
 		copyDtoToEntity(dto, entity);
 		entity = repository.save(entity);
 		return new ProductDTO(entity);
+		}
+		catch (EntityNotFoundException e) {
+			 throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
@@ -67,12 +79,17 @@ public class ProductService {
 	}
 	
 	
-	@Transactional
-	public void delete(Long id) 
-	{
-		repository.deleteById(id);
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
-	
-
 	
 }
